+++ src/io/io_result.h
@@ -1,6 +1,7 @@
 #ifndef IO_IO_RESULT_H_
 #define IO_IO_RESULT_H_
 
+#include <cstddef>
 #include <cstdint>
 
 namespace io {
@@ -16,23 +17,23 @@ enum class IoResultCode {
 class IoResult {
  public:
   static IoResult Pending();
-  static IoResult Read(uint64_t n);
+  static IoResult Read(size_t n);
   static IoResult Eof();
   static IoResult Error();
 
   bool ok() const { return code_ == IoResultCode::kOk; }
   bool pending() const { return code_ == IoResultCode::kPending; }
-  uint64_t nread() const { return nread_; }
+  size_t nread() const { return nread_; }
   bool eof() const { return code_ == IoResultCode::kEof; }
   bool error() const { return code_ == IoResultCode::kError; }
   IoResultCode code() const { return code_; }
 
  private:
   IoResult(IoResultCode code);
-  IoResult(uint64_t nread);
+  IoResult(size_t nread);
 
   IoResultCode code_;
-  uint64_t nread_;
+  size_t nread_;
 };
 
 }  // namespace io
